Move save, transforms into own files + organize
