+++ src/blocks/buttons/deprecated.js
@@ -6,7 +6,7 @@ import classnames from 'classnames';
 /**
  * Internal dependencies
  */
-import metadata from './../block.json';
+import metadata from './block.json';
 
 /**
  * WordPress dependencies
